extract validation test scenario
