+++ src/KeyHub.Integration.Tests/KeyHub.Integration.Tests.csproj
@@ -74,6 +74,7 @@
   <ItemGroup>
     <Compile Include="AccountTests.cs" />
     <Compile Include="CheckedDataContextFactory.cs" />
+    <Compile Include="LicenseValidationScenario.cs" />
     <Compile Include="LicenseValidatorTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TestSetup\CleanDatabaseAttribute.cs" />
